backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
slowapi>=0.1.9
requests>=2.31.0
groq>=0.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
rank-bm25>=0.2.2
openai>=1.40.0
google-genai>=1.0.0
biopython>=1.83
lxml>=5.0.0
Unidecode>=1.3.0
six>=1.16.0
